fix:生徒一覧のコントローラの統合
