+++ attendancemanagementsystem/src/main/java/com/example/attendancemanagementsystem/user/admin/controller/AdminStudentApiController.java
@@ -1,39 +1,39 @@
-package com.example.attendancemanagementsystem.user.admin.controller;
+// package com.example.attendancemanagementsystem.user.admin.controller;
 
-import java.util.Map;
+// import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.web.bind.annotation.GetMapping;
+// import org.springframework.web.bind.annotation.RequestMapping;
+// import org.springframework.web.bind.annotation.RequestParam;
+// import org.springframework.web.bind.annotation.RestController;
 
-import com.example.attendancemanagementsystem.user.admin.service.AdminStudentListService;
+// import com.example.attendancemanagementsystem.user.admin.service.AdminStudentListService;
 
-@RestController
-@RequestMapping("/admin/api")
-public class AdminStudentApiController {
+// @RestController
+// @RequestMapping("/admin/api")
+// public class AdminStudentApiController {
 
-    @Autowired
-    private AdminStudentListService adminStudentService;
+//     @Autowired
+//     private AdminStudentListService adminStudentService;
 
-    // 生徒一覧取得API
-    @GetMapping("/students")
-    public Map<String, Object> getStudents(
-            @RequestParam(name = "page", defaultValue = "0") int page,
-            @RequestParam(name = "size", defaultValue = "10") int size,
-            @RequestParam(name = "keyword", required = false) String keyword,
-            @RequestParam(name = "departmentId", required = false) Integer departmentId,
-            @RequestParam(name = "grade", required = false) Integer grade,
-            @RequestParam(name = "courseId", required = false) Integer courseId
-            ) {
+//     // 生徒一覧取得API
+//     @GetMapping("/students")
+//     public Map<String, Object> getStudents(
+//             @RequestParam(name = "page", defaultValue = "0") int page,
+//             @RequestParam(name = "size", defaultValue = "10") int size,
+//             @RequestParam(name = "keyword", required = false) String keyword,
+//             @RequestParam(name = "departmentId", required = false) Integer departmentId,
+//             @RequestParam(name = "grade", required = false) Integer grade,
+//             @RequestParam(name = "courseId", required = false) Integer courseId
+//             ) {
 
-        return adminStudentService.searchStudents(page, size, keyword, departmentId, grade, courseId);
-    }
+//         return adminStudentService.searchStudents(page, size, keyword, departmentId, grade, courseId);
+//     }
 
-    // フィルター選択肢取得API
-    @GetMapping("/search-options")
-    public Map<String, Object> getSearchOptions() {
-        return adminStudentService.getFilterOptions();
-    }
-}
+//     // フィルター選択肢取得API
+//     @GetMapping("/search-options")
+//     public Map<String, Object> getSearchOptions() {
+//         return adminStudentService.getFilterOptions();
+//     }
+// }
